main - move addon to main subfolder, split configs
