--- addons/grenadesintohatches/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\gih\addons\grenadesIntoHatches
