+++ addons/main/$PBOPREFIX$
@@ -1,0 +1,1 @@
+vik\gih\main
