+++ addons/main/XEH_preInit.sqf
