+++ addons/main/config.cpp
@@ -1,0 +1,21 @@
+#include "script_component.hpp"
+
+class CfgPatches {
+	class ADDON {
+		name = COMPONENT_NAME;
+		units[] = {};
+		weapons[] = {};
+		requiredVersion = REQUIRED_VERSION;
+		requiredAddons[] = {
+			"CBA_main"
+		};
+		author = MOD_AUTHOR;
+		url = MOD_URL;
+		version = VERSION;
+		versionStr = QUOTE(VERSION_STR);
+		versionAr[] = {QUOTE(VERSION_AR)};
+	};
+};
+
+#include "include/CfgFunctions.hpp"
+#include "include/CfgEventHandlers.hpp"
