+++ addons/main/functions/fn_applyCrewDamage.sqf
