+++ addons/main/functions/fn_applyVehicleDamage.sqf
