+++ addons/main/functions/fn_vehicleExplosion.sqf
@@ -5,7 +5,7 @@ params
 ];
 
 
-if (gih_getOutChanceEnabled) then {[_vehicle] call gih_fnc_crewChance};
+if (gih_getOutChanceEnabled) then {[_vehicle] call vik_gih_fnc_crewChance};
 
 _grenadeVisual = _grenadeAmmo createVehicle position _vehicle;
 // TODO - Find center of vehicle model
@@ -15,8 +15,8 @@ _grenadeVisual addEventHandler
 	"Deleted",
 	{
 		params ["_entity"];
-		[vehicle(attachedTo _entity)] spawn gih_fnc_applyCrewDamage;
-		[vehicle(attachedTo _entity)] spawn gih_fnc_applyVehicleDamage;
+		[vehicle(attachedTo _entity)] spawn vik_gih_fnc_applyCrewDamage;
+		[vehicle(attachedTo _entity)] spawn vik_gih_fnc_applyVehicleDamage;
 		_entity removeEventHandler [_thisEvent, _thisEventHandler];
 	}
 ];
