+++ addons/main/include/CfgFunctions.hpp
@@ -1,0 +1,14 @@
+class CfgFunctions
+{
+	class vik_gih
+	{
+		class functions
+		{
+            file = "vik\gih\main\functions";
+			class applyCrewDamage {};
+            class applyVehicleDamage {};
+            class crewChance {};
+            class vehicleExplosion {};
+		};
+	};
+};
