+++ addons/main/script_component.hpp
@@ -1,0 +1,4 @@
+#define COMPONENT main
+#define COMPONENT_BEAUTIFIED Main
+#include "\vik\gih\main\script_mod.hpp"
+#include "\vik\gih\main\script_macros.hpp"
